move new cli functions to config subcmd
